requirements.txt
azure-core==1.33.0
azure-storage-blob==12.25.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
filelock==3.18.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.2
greenlet==3.1.1
gunicorn==23.0.0
huggingface-hub==0.30.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pycparser==2.22
pyodbc==5.2.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==78.1.0
six==1.17.0
SQLAlchemy==2.0.40
sympy==1.13.1
tokenizers==0.21.1
tqdm==4.67.1
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
